Make Asciidoctor expose interfaces exclusively
